Create model for multi-cloud types
The normal Unet outputs the score for a pixel belonging to
a class.  The output has shape (batch_size, num_class, num_channel, height, width).
When there is just one cloud type, num_class=2:
0 for "not belonging to the cloud type"; 1 for "belonging to the cloud type".
So, if the score for 1 is greater than the score for 0, it means that
the model thinks this pixel is more likely to belong to the cloud type.

In this Kaggle challenge, there are four cloud types considered.  However,
it is noted that here, a pixel can belong to more than one cloud type, so
the problem is a multi-class multi-label classification of pixels.

In this commit, the usual last layer of a Unet is changed to contain four
convolutional layers, one for handling each cloud type.  So, the model output
consists of a tuple of four tensors of shape (batch_size, num_class,
num_channel, height, width).